evolve automata_tools
